+++ tests/Fixtures/RelCategory.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Ngos\AdminCore\Tests\Fixtures;
+
+use Illuminate\Database\Eloquent\Model;
+
+class RelCategory extends Model
+{
+    protected $table = 'rel_categories';
+
+    public $timestamps = false;
+
+    protected $guarded = [];
+}